Store refactor pre-commit
Up next:
The store will be refactored to support the new dispatcher. The
component initial state will be transformed to a model/dispatcher
containing the state structure.

The model prevents the dispatch method from changing the type during
runtime (it will log a warning and ignore the updated state).  In
addition the model has a (shallow) updated method that is exposed to
the store on render.

On each render the model is transformed to the next (upcoming) state
and every Object that not part of anan Array will be frozen using
Object#freeze, effectively making it immutable.

Summary:
The model will indicate the type of each state property, whether its
changed, prevents state mutation (besides Array elements), and will
enforce its structure on dispatch.